Overhauled audio system
Previously, pxt SFX, org waves, and org drums all had their own sample
structs and mixer functions. This is messy, and all mixer
implementations were buggy.

So now I've introduced a system that emulates DirectSound, reducing
redundancy and allowing the code to be much more accurate to the original.

On top of making the code itself more accurate, I also made Organya
behave more accurately: volume and panning should be accurate to
DirectSound's bizarre specification, and the frequency table now matches
the one in the Cave Story EXE, rather than the ones found in Org3.

All of this has also allowed me to make the volume balance accurate to
the original. The previous code was hackish, and only approximated the
behaviour of the original: notably, the org mixers manually halved the
samples because the volume algorithm was inaccurate to DirectSound.

In addition, I've reverted some of the changes between Org2 and Org3,
like the inclusion of playOrganObject2.